# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API de Señaventura",
    "version": "1.0.0",
    "description": ""
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    }
  ],
  "paths": {
...
    "/parteVideo/": {
      "get": {
        "tags": [
          "Parte Video"
        ],
        "description": "Endpoint para obtener todas las partes de video cuestionario.",
        "responses": {
          "200": {
            "description": "Partes de video cuestionario obtenidas correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id_parte_video_cuestionario": {
                        "type": "integer"
                      },
                      "id_nivel": {
                        "type": "integer"
                      },
                      "url_video": {
                        "type": "string"
                      },
                      "indice": {
                        "type": "integer"
                      },
                      "nombre": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener las partes de video cuestionario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Parte Video"
        ],
        "description": "Endpoint para crear una nueva parte de video cuestionario.",
        "responses": {
          "200": {
            "description": "Parte de video cuestionario creada correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_parte_video_cuestionario": {
                      "type": "integer"
                    },
                    "id_nivel": {
                      "type": "integer"
                    },
                    "url_video": {
                      "type": "string"
                    },
                    "indice": {
                      "type": "integer"
                    },
                    "nombre": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al crear la parte de video cuestionario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "id_nivel": {
                    "type": "integer"
                  },
                  "url_video": {
                    "type": "string"
                  },
                  "nombre": {
                    "type": "string"
                  }
                },
                "required": [
                  "id_nivel",
                  "url_video",
                  "nombre"
                ]
              }
            }
          }
        }
      }
    },
    "/parteVideo/byNivel/{id}": {
      "get": {
        "tags": [
          "Parte Video"
        ],
        "description": "Endpoint para obtener todas las partes de video cuestionario por nivel.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID del nivel."
          }
        ],
        "responses": {
          "200": {
            "description": "Partes de video cuestionario obtenidas correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id_parte_video_cuestionario": {
                        "type": "integer"
                      },
                      "id_nivel": {
                        "type": "integer"
                      },
                      "url_video": {
                        "type": "string"
                      },
                      "indice": {
                        "type": "integer"
                      },
                      "nombre": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener las partes de video cuestionario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/parteVideo/{id}": {
      "get": {
        "tags": [
          "Parte Video"
        ],
        "description": "Endpoint para obtener una parte de video cuestionario por su ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la parte de video cuestionario."
          }
        ],
        "responses": {
          "200": {
            "description": "Parte de video cuestionario obtenida correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_parte_video_cuestionario": {
                      "type": "integer"
                    },
                    "id_nivel": {
                      "type": "integer"
                    },
                    "url_video": {
                      "type": "string"
                    },
                    "indice": {
                      "type": "integer"
                    },
                    "nombre": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al obtener la parte de video cuestionario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Parte Video"
        ],
        "description": "Endpoint para eliminar una parte de video cuestionario.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la parte de video cuestionario a eliminar."
          }
        ],
        "responses": {
          "200": {
            "description": "Parte de video cuestionario eliminada correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_parte_video_cuestionario": {
                      "type": "integer"
                    },
                    "id_nivel": {
                      "type": "integer"
                    },
                    "url_video": {
                      "type": "string"
                    },
                    "indice": {
                      "type": "integer"
                    },
                    "nombre": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al eliminar la parte de video cuestionario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "required": false,
          "content": {}
        }
      },
      "put": {
        "tags": [
          "Parte Video"
        ],
        "description": "Endpoint para editar una parte de video cuestionario.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID de la parte de video cuestionario a editar."
          }
        ],
        "responses": {
          "200": {
            "description": "Parte de video cuestionario editada correctamente.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_parte_video_cuestionario": {
                      "type": "integer"
                    },
                    "id_nivel": {
                      "type": "integer"
                    },
                    "url_video": {
                      "type": "string"
                    },
                    "indice": {
                      "type": "integer"
                    },
                    "nombre": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error al editar la parte de video cuestionario.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "url_video": {
                    "type": "string"
                  },
                  "nombre": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}